feat(DEV-1552): move helpers functions
